Add config option to select metadata source (Theme/ES)
This theme is best viewed when the system names does not exceed some
length. For this reason, this theme comes with adjusted metadata
containing the system name to display (e.g. "Family Computer Disk
System" (ES) -> "Famicom Disk System" (Theme)). These metadata also
include the release year and controller to display.

We do not control the ES systems metadta (which comes from
"es_systems.cfg"). Therefore, providing the metadata from the theme
ensure the best user experience out of the box.

However, the downside of using theme metadata is that it can cause
inconsistencies with some ES features. For example, systems can be
sorted by release year. If the release year of a system is different
between ES and the Theme, then the displayed year will not look correct
in regard to the sorted systems.

The solution for the best of both worlds is to override the ES metadata
with the theme ones. Then, we can configure the Theme to use ES as
source. As a result, displayed system names will be appropriate, and
also these values will be used internally by ES, ensuring consistency.

This is reserved to advanced user. I think inconsitencies between
metadata should not happen often enough for casual user to notice. Yet,
I'll later provide a Python util script in charge of updating ES config
with Theme metadata.