Simplify and merge SVG / PNG for system carousel background
What a journey...

I wanted to combine the background and shadow of the carousel (currently
being two different files) into a single image to simplify the placement
in the theme.

By doingo so, it also makes the SVG code of the background much simpler
and more readable (hand-crafted instead of software output). Writing SVG
is not unpleasant in itself, except when you realize how poorly and
disparately SVG files are supported.

In particular, EmulationStation seems to use a very old and simplist
SVG render engine. Some SVG features did not work at all, maybe that's
why the shadow was a separate file.

A preferable solution, then, is to use PNG instead of SVG (which is also
apparently better for performance). But then, again, the conversion
isn't straightforward and not all tools support all features despite how
common SVG files are! I encountered invalid rendering with both
imagemagick and inkscape. Therefore, I had to rely on a canvas rendering
the SVG in the browser in order to then being able to save the image as
a PNG manually.

The carousel background is now described as a 1920x1080 image for
simplicity's sake.

The borders of the rectangle look maybe a little too sharp compared to
the original. I tried to apply a subtle blur filter to the rectangle,
but it generated undesirable white artefacts, therefore I gave up.