Trim all system logo image to remove transparent borders
Previously, the images were normalized to 540x540 resolution and logo
were centered with some margin. I want to experiment with new carousel
dimensions (restricted to avoid overlap with bottom bar), but I'd like
to keep most of the logo to the same size as before visually. To do so,
I need to know the true size of the logo, to optimally place them on the
UI without the transparent borders that consumes virtual space.

This commit can always be reverted if needed.

Command used for cropping:
    for a in *.webp; do convert -trim "$a" "$a"; done