Improve quality of "Playdate" logo
Quality was visually improved by actually reverting latest lossless
compression and restauring the logo to the old (lossy) one.

For an unknown reason, the logo was of objectively better quality, but
once rendered in ES it generated small undesirable artefacts.

The "Playdate" logo is a bit special because this is the only one that I
created, the other ones were copied from other ES themes. Therefore,
it's very likely the others logos were already resized and encoded
lossily. Maybe a "smooth" effect was applied during this operation,
which counterbalances the ES artefacts visible on images with more
sharpness.