Trim all system controller images to remove transparent borders
See related previous commits. This makes it easier to reposition and
resize the image on the UI without the risk of overlap with
other elements.

This time, I discovered yet another quirk: using "<maxSize>" sometimes
causes ES to generate very weird artefacts around the image. This
depends on the value chosen for resizing.

What's incomprehensible is that if you use "<size>" with a
constraint for a single dimension (width), then there's no problem (see
the "system-carousel.xml" diff for more tests). No artifacts
generated when the image output dimensions are the same (just display
the boxes to see for yourself).

Another thing: deactivating the "Optimize Images for Vram usage"
option also corrects the problem. So that's an option I'll always
disable.