Add option to enable/disable "<linearSmooth>"
When turned off, the resized images of system controllers sometimes have
artefacts for an unknown reason (likely a bug in ES). This is way we
need to turn it on by default. However, the images are blurry in this
case. This can be solved by disabling smooth resizing, but to avoid the
artefacts bug, the VRAM optimizations must also be disabled.

See also: 537e57216df70a317f29d23d2b7f7a765858d650