Fix some systems not recognized by Batocera due to misnaming
It seems ES and ES-DE conventions differ on some of the system expected
names.

We should think about some kind of pseudo-symlinks in the future, just
to make sure we cover as many implementations as possible (since the
"es_systems.cfg" file can be created by anyone, not just Batocera).